# APICL.Client/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "APICL.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/Audio/audios": {
      "get": {
        "tags": [
          "Audio"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/AudioObjInfo"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Audio/audios/{guid}/info": {
      "get": {
        "tags": [
          "Audio"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guid",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AudioObjInfo"
            }
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Audio/audios/{guid}/remove": {
      "delete": {
        "tags": [
          "Audio"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guid",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "boolean"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Audio/audios/upload": {
      "post": {
        "tags": [
          "Audio"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "copyGuid",
            "type": "boolean",
            "default": true
          },
          {
            "in": "formData",
            "name": "file",
            "type": "file"
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/AudioObjInfo"
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Audio/audios/{guid}/download": {
      "get": {
        "tags": [
          "Audio"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guid",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Audio/audios/{guid}/waveform64": {
      "get": {
        "tags": [
          "Audio"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guid",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AudioData"
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Audio/audios/{guid}/play/{volume}": {
      "post": {
        "tags": [
          "Audio"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guid",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "volume",
            "required": true,
            "type": "number",
            "format": "float",
            "default": 0.66
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Audio/audios/{guid}/stop": {
      "post": {
        "tags": [
          "Audio"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guid",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AudioObjInfo"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Audio/audios/stopAll": {
      "post": {
        "tags": [
          "Audio"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AudioData": {
      "type": "object",
      "properties": {
        "guid": {
          "format": "uuid",
          "type": "string"
        },
        "filepath": {
          "type": "string"
        },
        "length": {
          "format": "int64",
          "type": "integer"
        },
        "pointer": {
          "format": "int64",
          "type": "integer"
        },
        "waveformBase64": {
          "type": "string"
        },
        "waveformWidth": {
          "format": "int32",
          "type": "integer"
        },
        "waveformHeight": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "AudioObjInfo": {
      "type": "object",
      "properties": {
        "guid": {
          "format": "uuid",
          "type": "string"
        },
        "filepath": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "samplerate": {
          "format": "int32",
          "type": "integer"
        },
        "bitdepth": {
          "format": "int32",
          "type": "integer"
        },
        "channels": {
          "format": "int32",
          "type": "integer"
        },
        "length": {
          "format": "int32",
          "type": "integer"
        },
        "memoryLocation": {
          "type": "string"
        },
        "chunkSize": {
          "format": "int32",
          "type": "integer"
        },
        "overlapSize": {
          "format": "int32",
          "type": "integer"
        },
        "form": {
          "type": "string"
        },
        "bpm": {
          "format": "float",
          "type": "number"
        },
        "playingString": {
          "type": "string"
        },
        "sizeString": {
          "type": "string"
        },
        "durationTimeSpan": {
          "format": "date-span",
          "type": "string",
          "readOnly": true
        },
        "durationString": {
          "type": "string"
        },
        "lastProcessingTime": {
          "type": "string"
        },
        "lastLoadingTime": {
          "type": "string"
        },
        "entry": {
          "type": "string"
        },
        "errorInfo": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    }
  }
}